+++ Changelog.md
@@ -3,7 +3,7 @@
 ## 1.0.1
 
  - A `factorial()` method has been added to `BinaryInteger`
- - The `@Transformed` property wrapper has been added
+ - Added `@Transformed` and `@Validated` property wrappers
 
 ## 1.0.0